Refactor LimaDetector
Refactor the class LimaDetector to include the tango testing layer.
Create the LimaCoreDector to test the detecto by using the Lima core
objects.

Move the acquisition test logic to the LimaCCDAcquistiionTest class,
this allows to use the same logic to test the detector by using the
Lima Core and the Tango Layer.